// packages/backend/src/types/schemas/prisma-generated/inputTypeSchemas/InvitationCodeCreateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { UserCreateNestedOneWithoutCreatedInvitationsInputSchema } from './UserCreateNestedOneWithoutCreatedInvitationsInputSchema';
import { UserCreateNestedOneWithoutUsedInvitationsInputSchema } from './UserCreateNestedOneWithoutUsedInvitationsInputSchema';

export const InvitationCodeCreateInputSchema: z.ZodType<Prisma.InvitationCodeCreateInput> = z.object({
  id: z.string().optional(),
  code: z.string(),
  expiresAt: z.coerce.date(),
  isUsed: z.boolean().optional(),
  isActive: z.boolean().optional(),
  maxUses: z.number().int().optional(),
  currentUses: z.number().int().optional(),
  createdAt: z.coerce.date().optional(),
  updatedAt: z.coerce.date().optional(),
  createdBy: z.lazy(() => UserCreateNestedOneWithoutCreatedInvitationsInputSchema),
  usedBy: z.lazy(() => UserCreateNestedOneWithoutUsedInvitationsInputSchema).optional()
}).strict();

export default InvitationCodeCreateInputSchema;